Move fmtDayLabel, fmtTime, toDayKey into fmtDate utility
Eliminates duplication between Timeline.tsx and LogPrintSheet.tsx.
